Added conveyor and placed into the map

+++ core/src/main/java/io/github/eco_warrior/constant/ConstantsVar.java
@@ -2,8 +2,8 @@ package io.github.eco_warrior.constant;
 
 public class ConstantsVar {
 
-    public static final int WINDOW_HEIGHT = 720;
-    public static final int WINDOW_WIDTH = 1280;
+    public static final float WINDOW_HEIGHT = 720;
+    public static final float WINDOW_WIDTH = 1280;
     public static final String WINDOW_TITLE = "Eco Warrior";
 
 }
